+++ orthanc_api_client/__init__.py
@@ -4,6 +4,7 @@ from .helpers import *
 from .change import ChangeType, ResourceType
 from .study import Study, StudyInfo
 from .series import Series, SeriesInfo
+from .instances_set import InstancesSet
 from .job import Job, JobInfo, JobType, JobStatus
 from .http_client import HttpClient
 from .downloaded_instance import DownloadedInstance