+++ orthanc_api_client/series.py
@@ -1,5 +1,5 @@
 from .tags import SimplifiedTags
-import typing
+from typing import List, Optional
 
 
 class SeriesInfo:
@@ -12,14 +12,24 @@ class SeriesInfo:
         self.study_orthanc_id = json_series.get('ParentStudy')
 
 
+class SeriesStatistics:
+
+    def __init__(self, json_series_stats):
+        self.instances_count = int(json_series_stats['CountInstances'])
+        self.disk_size = int(json_series_stats['DiskSize'])                  # this is the total size used on disk by the series and all its attachments
+        self.uncompressed_size = int(json_series_stats['UncompressedSize'])  # this is the total size of the series and all its attachments once uncompressed
+
+
 class Series:
 
-    def __init__(self, api_client, orthanc_id):
+
+    def __init__(self, api_client: 'OrthancApiClient', orthanc_id: str):
         self._api_client = api_client
         self.orthanc_id = orthanc_id
-        self._info = None
-        self._study = None
-        self._instances = None
+        self._info: SeriesInfo = None
+        self._statistics: SeriesStatistics = None
+        self._instances: Optional[List['Instances']] = None
+        self._study: Optional['Study'] = None
 
     @staticmethod
     def from_json(api_client, json_series: object):
@@ -30,7 +40,8 @@ class Series:
     @property
     def info(self):  # lazy loading of main dicom tags ....
         if self._info is None:
-            self._load_info()
+            json_series = self._api_client.series.get_json(self.orthanc_id)
+            self._info = SeriesInfo(json_series)
         return self._info
 
     @property
@@ -41,9 +52,12 @@ class Series:
     def dicom_id(self):
         return self.info.dicom_id
 
-    def _load_info(self):
-        json_series = self._api_client.series.get_json(self.orthanc_id)
-        self._info = SeriesInfo(json_series)
+    @property
+    def statistics(self):  # lazy loading of statistics ....
+        if self._statistics is None:
+            json_series_stats = self._orthancClient.series.get_json_statistics(self.id)
+            self._statistics = SeriesStatistics(json_series_stats)
+        return self._statistics
 
     @property
     def study(self) -> 'Study':  # lazy creation of study object
@@ -52,7 +66,7 @@ class Series:
         return self._study
 
     @property
-    def instances(self) -> typing.List['Instance']:  # lazy creation of instances objects
+    def instances(self) -> List['Instance']:  # lazy creation of instances objects
         if self._instances is None:
             self._instances = []
             for instance_id in self.info.instances_orthanc_ids: