+++ release-notes.md
@@ -3,6 +3,7 @@ v 0.9.0
 
 - **BREAKING CHANGE:** renamed `download_study` and `download_series` into `download_instances`
 - introduced `Series`, `SeriesInfo`, `Instance` and `InstanceInfo` classes
+- introduced `InstancesSet` class
 
 v 0.8.3
 =======